requirements.txt
anthropic==0.34.4
slack-sdk==4.27.2
sqlalchemy==2.1.5
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.3.4
gitpython==3.1.43
aiosqlite==3.2.1
asyncio-contextmanager==1.0.0
psutil==6.0.0
requests==2.32.3
